refactor(transfer-accounts): Move TransferAccounts to Movements namespace
